+++ packages/ai-constructs/src/conversation/runtime/event-tools-provider/graphql_query_factory.ts
@@ -15,12 +15,11 @@ export class GraphQlQueryFactory {
     const { graphqlRequestInputDescriptor } = toolDefinition;
     const { selectionSet, queryName } = graphqlRequestInputDescriptor;
     const [topLevelQueryArgs, queryArgs] = this.createQueryArgs(toolDefinition);
-
+    const fieldSelection =
+      selectionSet.length > 0 ? ` { ${selectionSet} }` : '';
     const query = `
     query ToolQuery${topLevelQueryArgs} {
-      ${queryName}${queryArgs} {
-        ${selectionSet}
-      }
+      ${queryName}${queryArgs}${fieldSelection}
     }
   `;
 
@@ -36,7 +35,20 @@ export class GraphQlQueryFactory {
     }
 
     const { properties } = inputSchema.json as InputSchemaJson;
-    if (!properties) {
+
+    // The conversation resolver should not pass an empty object as input,
+    // but we're defensively checking for it here anyway because if `properties: {}`
+    // is passed, it will generate invalid GraphQL. e.g.
+    // Valid:
+    // query ToolQuery {
+    //   exampleQuery
+    // }
+    //
+    // Invalid:
+    // query ToolQuery {
+    //   exampleQuery()
+    // }
+    if (!properties || Object.keys(properties).length === 0) {
       return ['', ''];
     }
     const { propertyTypes } = toolDefinition.graphqlRequestInputDescriptor;