+++ packages/backend-secret/src/ssm_secret_with_amplify_error_handling.ts
@@ -81,6 +81,7 @@ export class SSMSecretClientWithAmplifyErrorHandling implements SecretClient {
           'AccessDeniedException',
           'NotAuthorized',
           'ExpiredTokenException',
+          'ExpiredToken',
           'CredentialsProviderError',
           'InvalidSignatureException',
         ].includes(error.cause.name)