+++ packages/backend-ai/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-ai
 
+## 0.1.1
+
+### Patch Changes
+
+- c85d071: updated packages to fix violations of prefer-amplify-errors rule
+- Updated dependencies [8e964e1]
+  - @aws-amplify/ai-constructs@0.1.4
+
 ## 0.1.0
 
 ### Minor Changes